sistema_spec/requirements.txt
PySide6
bcrypt
requests
pandas
openpyxl
fpdf2